requirements.txt
appnope==0.1.2
backcall==0.2.0
blinker==1.4
click==7.1.2
decorator==4.4.2
Flask==1.1.2
Flask-Login==0.5.0
flask-requests==0.0.14
Flask-SQLAlchemy==2.4.4
greenlet==1.0.0
importlib-metadata==3.7.3
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
jupyter-client==6.1.12
jupyter-core==4.7.1
MarkupSafe==1.1.1
pampy==0.3.0
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.17
ptyprocess==0.7.0
Pygments==2.8.1
python-dateutil==2.8.1
pyzmq==22.0.3
six==1.15.0
SQLAlchemy==1.4.0
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
wcwidth==0.2.5
Werkzeug==1.0.1
zipp==3.4.1
